+++ src/views/join/components/index.ts
