+++ src/views/join/components/info.tsx
