+++ src/views/join/components/join-stepper.tsx
