+++ src/views/join/components/term-form.tsx
