+++ src/views/join/sections/index.ts
@@ -1,0 +1,3 @@
+export * from './join-form-section'
+export * from './join-success-section'
+export * from './join-term-section'
